requirements.txt
# Core
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Data
yfinance>=0.2.40

# Deep learning (CPU build is fine for most users)
tensorflow-cpu>=2.15.0

# Viz & Utils
matplotlib>=3.7.0
seaborn>=0.12.2

# Optional (for notebooks)
jupyter>=1.0.0
ipykernel>=6.0.0
